Not-understood change for galaxy player
From doing a wireshark USB comparison, the end file transfer packet sent
by Odin has a 2 for the unknown2 field, and heimdall was using a 0.
Changing this to a 2 makes heimdall work for the Galaxy Player 5.0. Note
that I have no idea what this field is or what the consequences of
changing it to a 2 are!